requirements.txt
setuptools>=60.5.0
asyncio>=3.4.3
TemporaryStorage>=2022.1.31.2
Pillow>=8.4.0